chore: moved AnalyticsService protocol implementation into Firebase and Segment managers
